requirements.txt
Flask>=3.0.0
PyPDF2>=3.0.1
pandas>=2.0.0
openpyxl>=3.1.0
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
opencv-python>=4.8.0
pyttsx3>=2.90
gTTS>=2.4.0
